Before we could add a product into the cart or increment if it was already added, we had to check if it was already in the cart and increase it's quantity by one if it was or just add if if the cart was empty.
"FindProductInCartLines" class was pre-created, and a foreach loop was implemented in a way it could get "GetCartLineList" returned and check each cartItem one by one comparing them with productId.

If the result was a bool true implying there's a product, it will return the product in the cart.

Else, it will return null if nothing is found.(exception)

Then we can go back to the AddItem method and return the product in the cart line if it is found, there.

We instanciate CartLine, creating a new CartLine because we need to add it to a List of CartLine.