AddItemInCart was fixed.
On class AddItem if "simple" is not null, means an item was found in the cart, then, foreach loop comes into action, going through each product of cartline list, and, the ones who are in cartLine are then compared by their product.Id - If they are the same, add this item's quantity in cartLine by 1

Else, just add the cartItem to the cart.

Now the "null" comes from class FindProductInCartLines returns "null". Such happens if the product was not found in the cart.

Here the foreach loop iterates through each cartItem in GetCartLineList and if a product is found it will be returned. Only when the product is not found in GetCartLineList, null will be the result enabling us to add cartItem to the cart.

Otherwise, the result will be different from null, activating the cartLine.Product.Id == product.Id comparison, which will increase it's quantity by 1(not add!).

Moving on to next test.