Here GetAverageValue was fixed.
To sum up:

a double variable was created to hold the totalValueOfCart.

We iterated through each item int he cartline list.

We made totalValueOfCart the sum of itself times the cartItem Quantity.

We used a lambda expression to summarize the sum of items to count from the returned cartline list.

we devided the totalValueOfCart by countItemsInCartLine and got the average value.

we can now return the average value of all items in the cart by calling GetAverageValue();