SetCulture and GetProductById were fixed.
Regarding SetCulture, a switch statement was used, code was commented for further detail.

While doing GetProductById, GetAllProducts() was a method with a return type of array and code was refactored to List<t>. It was working without such change, but was done as requested, through Implementation a nd Definition.

On GetProductById, similar principle used in AddItem:

Aforeach loop comparing product.Id with the passed id through the method GetProductById, returning the product if there is a match "==" and if there isn't, return null, which will result in the original error, because the id doesn't exist.